move user interfaces and dtos to @shared
