+++ apps/backend/src/validators/login.validator.ts
@@ -1,9 +1,6 @@
 import Joi from "joi";
 
-export interface LoginDto {
-  password: string;
-  username: string;
-}
+import { LoginDto } from "@shared/dtos/login.dto";
 
 export const validate = (payload: any): Promise<LoginDto> =>
   Joi.object<LoginDto>({
