+++ apps/shared/dtos/login.dto.ts
@@ -1,0 +1,4 @@
+export interface LoginDto {
+  password: string;
+  username: string;
+}
