+++ apps/shared/dtos/register.dto.ts
@@ -1,0 +1,4 @@
+export interface RegisterDto {
+  password: string;
+  username: string;
+}
