+++ apps/shared/interfaces/user.ts
@@ -1,0 +1,3 @@
+export interface User {
+  username: string;
+}
